+++ windows/flutter_local_authentication_plugin.h
@@ -4,7 +4,9 @@
 #include <flutter/method_channel.h>
 #include <flutter/plugin_registrar_windows.h>
 
+#include <functional>
 #include <memory>
+#include <string>
 
 namespace flutter_local_authentication {
 
@@ -24,6 +26,12 @@ class FlutterLocalAuthenticationPlugin : public flutter::Plugin {
   void HandleMethodCall(
       const flutter::MethodCall<flutter::EncodableValue> &method_call,
       std::unique_ptr<flutter::MethodResult<flutter::EncodableValue>> result);
+
+ private:
+  HWND window_handle_ = nullptr;
+  std::function<void()> cancel_authentication_;
+  bool cancellation_requested_ = false;
+  std::string localization_reason_ = "Authentication is required.";
 };
 
 }  // namespace flutter_local_authentication