+++ windows/test/flutter_local_authentication_plugin_test.cpp
@@ -39,5 +39,33 @@ TEST(FlutterLocalAuthenticationPlugin, GetPlatformVersion) {
   EXPECT_TRUE(result_string.rfind("Windows ", 0) == 0);
 }
 
+TEST(FlutterLocalAuthenticationPlugin, ReportsWindowsHelloBiometryType) {
+  FlutterLocalAuthenticationPlugin plugin;
+  std::string biometry_type;
+  plugin.HandleMethodCall(
+      MethodCall("getBiometryType", std::make_unique<EncodableValue>()),
+      std::make_unique<MethodResultFunctions<>>(
+          [&biometry_type](const EncodableValue* result) {
+            biometry_type = std::get<std::string>(*result);
+          },
+          nullptr, nullptr));
+
+  EXPECT_EQ(biometry_type, "multiple");
+}
+
+TEST(FlutterLocalAuthenticationPlugin, CancelWithoutAuthenticationReturnsFalse) {
+  FlutterLocalAuthenticationPlugin plugin;
+  bool canceled = true;
+  plugin.HandleMethodCall(
+      MethodCall("cancelAuthentication", std::make_unique<EncodableValue>()),
+      std::make_unique<MethodResultFunctions<>>(
+          [&canceled](const EncodableValue* result) {
+            canceled = std::get<bool>(*result);
+          },
+          nullptr, nullptr));
+
+  EXPECT_FALSE(canceled);
+}
+
 }  // namespace test
 }  // namespace flutter_local_authentication